% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-data-frame.R
\name{create_data_frame_in_c}
\alias{create_data_frame_in_c}
\title{Create a data.frame within C code and return to R}
\usage{
create_data_frame_in_c()
}
\description{
Create a data.frame within C code and return to R
}
